fix(windows): fix build issues (#36)
* deps(windows): add missing features

On Windows, the crossterm crate requires the "windows" feature.
Additionally, the windows crate (or one of its optional features, I'm
not sure which) requires the "std" feature, and the winsafe crate needed
the "advapi" feature (for GetUserName()) and "kernel" (for
GlobalMemoryStatusEx()) features.

* fix(windows): wininfo::memory() failed to build

winsafe's GlobalMemoryStatusEx() does not take any parameters, but
simply returns the memory information in a SysResult<MEMORYSTATUSEX>.  I
presume the call with a parameter was an artifact of a previous
implementation that used the `windows` crate equivalent, which *does*
take a mutable struct reference parameter to pass the result back.