Respect /etc/machine-info overrides on Linux (#34)
* refactor(cleanup): Split out the parser logic from `os_release` method

Multiple configuration files specified by systemd are specified as "a
newline-separated list of environment-like shell-compatible variable
assignments" (os-release, machine-info, locale.conf, etc).

* fix(linux): Don't spit out a failure if the "env-like shell-like" file does not exist

If the file does not exist, then treat it as if it were empty and therefore
had no assignments in it.

This generally wasn't a problem for /etc/os-release (nearly present in every
distro) but does come up for other files we'd want to parse this way.

* fix(linux): use admin-configured model/vendor name from /etc/machine-info if present

This allows a user to provide a meaningful model name if their system vendor
left it as `To Be Filled By O.E.M.` or similar.

Fixes #33, for the Linux case.